% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samplingfuns.R
\name{apcsamp}
\alias{apcsamp}
\title{A wrapper for "draw_chains" that will use multiple search chains in the algorithm.}
\usage{
apcsamp(dat, dv = "y", apc = c("a", "p", "c"), cores = 1,
  method = "gibbs", chains = 1, samples = 100, draws = 1000)
}
\arguments{
\item{dat}{A dataframe.}

\item{dv}{A character variable or column number indicating the dependent variable.}

\item{apc}{A character vector naming the age, period, and cohort variables (defaults to 'a', 'p', and 'c').}

\item{cores}{Integer indicating cores to use.}

\item{method}{One of "gibbs" or "ml". For a fully bayesian estimator for a faster approximation.}

\item{samples}{Integer indicating the number of models to sample in the chain.}

\item{draws}{Integer indicating samples; used only for "gibbs" method.}

\item{startvals}{Optional list for chain starting values.}
}
\value{
An object of apcsamp.
}
\description{
A wrapper for "draw_chains" that will use multiple search chains in the algorithm.
}
\details{
what does this look like.
}
\examples{
data(apcsim)
apcsamp(apcsim,method='ml',samples=250,cores=4,chains=4)

}
