% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samplingfuns.R
\name{draw_effs}
\alias{draw_effs}
\title{Draws effect estimates from an APC sample using Bayesian Model Averaging.}
\usage{
draw_effs(sampobj, means = FALSE, betas = FALSE, tol = NULL)
}
\arguments{
\item{sampobj}{An apcsamp object (\code{\link{apcsamp}})}

\item{means}{Logical: TRUE returns mean estimates; FALSE generates centered estimates (relatvie to grand mean).}

\item{betas}{Logical: TRUE returns actual estimates.}

\item{tol}{A number between 0 and 1, which indicates tolerance to rounding and the number
to be drawn. e.g. tol=0.001 is tolerance to the third decimal and draws 1,000 samples.}
}
\value{
An object of apceffects.
}
\description{
Draws effect estimates from an APC sample using Bayesian Model Averaging.
}
\examples{
data(apcsim)

sampleobject = apcsamp(apcsim,method='ml',
samples=250,cores=4,chains=4)

effectsobject = draw_effs(sampleobject)
}
