% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effectsoperations.R
\name{modeled_windows}
\alias{modeled_windows}
\title{Returns a dataframe with factor variables including window breaks at a specifie level.}
\usage{
modeled_windows(effectsobj, pval = 0.05)
}
\arguments{
\item{effectsobj:}{an apc effects object.}

\item{pval:}{a probability-value indicating a cut-off for the window breaks}
}
\value{
A dataframe with factor varaibles appended ".win" based on the breaks.
}
\description{
Returns a dataframe with factor variables including window breaks at a specifie level.
}
