% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samplingfuns.R
\name{postsamp}
\alias{postsamp}
\title{A wrapper for "draw_effs" that will use multiple cores to speed up sampling from the posterior.}
\usage{
postsamp(sampobj, tol = NULL, cores = NULL, ...)
}
\arguments{
\item{sampobj}{A sample object.}
}
\value{
An object of apceffects.
}
\description{
A wrapper for "draw_effs" that will use multiple cores to speed up sampling from the posterior.
}
