% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samplingfuns.R
\name{swslm}
\alias{swslm}
\title{A wrapper that samples from model space and draws from posterior.}
\usage{
swslm(formula, ...)
}
\arguments{
\item{formula}{a formula.}
}
\value{
An object of apceffects.
}
\description{
A wrapper that samples from model space and draws from posterior.
}
