% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effectsoperations.R
\name{tstdiff}
\alias{tstdiff}
\title{Test for statistical differneces in window breaks.}
\usage{
tstdiff(effectsobj)
}
\arguments{
\item{effectsobj}{An apceffects object.}
}
\value{
Provides a list for each winow object showing whether there are statistical differences.
}
\description{
Test for statistical differneces in window breaks.
}
