% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samplingfuns.R
\name{draw_chains}
\alias{draw_chains}
\title{Draw a single chain.}
\usage{
draw_chains(formula, data, windowvars = c("a", "p", "c"), cores = 1,
  method = "ml", chains = 1, samples = 10, draws = 1000,
  startvals = NULL)
}
\arguments{
\item{formula}{A formula object.}

\item{data}{A dataframe.}

\item{windowvars}{A character vector naming the age, period, and cohort variables (defaults to 'a', 'p', and 'c'). This can also be used to calculate breaks for other sorts of variables.}

\item{cores}{Integer indicating cores to use.}

\item{method}{One of "gibbs" or "ml". For a fully bayesian estimator for a faster approximation.}

\item{samples}{Integer indicating samples; used only for "gibbs" method.}

\item{draws}{Integer indicating the number of models to sample in the chain.}

\item{startvals}{Optional list for chain starting values.}
}
\value{
An object of apcsamp.
}
\description{
Draw a single chain.
}
\examples{
data(apcsamp)
draw_chains(apcsamp,method='ml',samples=250,cores=4,chains=4)
}
